Implement duke package
