+++ data/duke.txt
@@ -1,0 +1,1 @@
+T | 1 | blah
