+++ src/main/java/Duke.java
@@ -1,3 +1,5 @@
+import duke.*;
+
 public class Duke {
 
     private Storage storage;
