+++ src/main/java/duke/DeleteCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DeleteCommand extends Command {
 
     private int taskNumber;
@@ -7,9 +9,10 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         Task deletedTask = tasks.delete(taskNumber);
         ui.showDeletedTask(deletedTask, tasks);
+        tasks.updateStorage(storage);
     }
 
 }
