+++ src/main/java/duke/ToDo.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Encapsulate Todo which is-a Task.
  *
