+++ src/main/java/duke/UnmarkCommand.java
@@ -1,14 +1,17 @@
-public class MarkCommand extends Command {
+package duke;
+
+public class UnmarkCommand extends Command {
 
     private int taskNumber;
 
-    public MarkCommand(int taskNumber) {
+    public UnmarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
+
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        Task markedTask = tasks.mark(taskNumber);
-        ui.showMarkedTask(markedTask);
+        Task unmarkedTask = tasks.unmark(taskNumber);
+        ui.showUnmarkedTask(unmarkedTask);
         tasks.updateStorage(storage);
     }
 }
